Extract sessionWindowStrategy IT to IoTDBUDFSessionWindowQueryIT for avoiding LocalStandaloneIT
